Add method to dynamically bind method context
This is a challenge for resources that export values they don't control.
If that export depends on a stable `this` identity, let's say a private
field, the method will blow up in an unexpected way because the `this`
context wasn't bound.

This commit adds a utility proxy for doing just that.