Rename package
bobcat -> wardens

Bobcat was more of a codename, something unique to grep for until I got
a better name. "Wardens" represents a keeper, custodian, administrator,
and an idea of a prison where things are kept under control. It's not
a great name, but with the state of name availability on npm, it's
pretty good.