Only require config if the class expects it
`mount(...)` used to require a second parameter even if the class didn't
have an `enter()` handler, or if it didn't accept a config. Now
`mount(...)` is smart about it. It also accepts variable arguments now.

I'm not stoked about variable arguments, I'd rather see context objects,
but every other attempt with TypeScript to represent an optional
argument conditional on a generic resulted in significant pain. At least
variable args are on the happy path.