Define mount/unmount interfaces
Obviously I can't provide a concrete type for `enter(...)` in the
abstract class, but I didn't want to require it the way I did with
`exports`. I used the generic to accomplish that, but I think I can
further improve the experience.

I've added interfaces for `enter`/`leave`. Now I don't need an
implementation in the superclass, and TypeScript can infer it from the
concrete definition.