Replace class with functional approach
This commit entirely replaces the `Resource` class with a functional
approach. You write an async function and it returns a resource.

The main reason for this change is avoiding null conditions: classes
must be instantiated synchronously, but resources are allocated
asynchronously. Putting the `create` handler on the class necessitates
non-null assertions (`value!: SomeValue`) that can't be statically
proven. By coupling to the class, there's no way around this.

By moving to async functions as the `create` handler, the null case is
avoided and you can still use classes as a return value:

```
const Demo = async () => ({
  value: new Controls(await getValue()),
})
```

This also has a happy side effect of consistent naming. I couldn't name
`allocate`/`deallocate` the same thing as the handlers because they
shared a namespace. I had to invent new names. Now it's a parameter so
all the naming can be consistent.