Clean up orphaned resources under case of failure
This addresses an edge case: if a resource creates a child resource and
then fails halfway through the rest of its setup, the child resource is
leaked.

Now `create(...)` detects allocation failures and will attempt to
destroy all the intermediate (orphaned) resources.

This also improves error reporting when 2 or more child resources cannot
be destroyed.