Add context API
The context API is inspired by React's context, although I entertained
a few alternatives before settling here. The pain point is that passing
state from an grandparent resource to a child resource requires adding
new parameters the whole way down, creating a lot of noise and useless
plumbing.

I'd like to be able to transfer values to different parts of the tree
without needing to rewrite every resource and unit test to support a new
value.

The context API pulls state from a prototype chain which mirrors the
resource hierarchy. Setting or overriding a value on the chain makes it
available to your own resource and anything you create down the chain.