Restructure files to better represent roles
I cut quite a few corners in the naming and organization of the project.
Figured more meaningful names would be easier to assign later in the
process. I was right.

Most of this is just renaming files, but I also renamed
`ResourceContext` to avoid ambiguity with the new context API.