Only allow destroy to tear down root resources
I ran into this accidentally in one of my projects. I refactored
a resource to return a child resource (an HTTP server) and ran a test on
it, then destroyed the resource. But during the refactor I failed to
notice that the parent resource was still around. I would have expected
some sort of error telling me that I don't own the resource, and the
parent has to be cleaned up too.

This adds an error case to make sure you can only destroy resources you
created.