Automate the release process
This is adapted from another of my projects, `signals`, which publishes
RC versions automatically when merging to main and has a manual dispatch
workflow for promoting a stable version.

Publishing is incredibly error prone. I constantly make mistakes, such
as using a non-annotated git tag, copying the wrong changelog revision,
or forgetting the rc tag on npm publish.

The hope is this makes releases more robust.